fixed: add Code to Profile page , delete unuse file
